+++ core/flamingock-test-support/src/main/java/io/flamingock/support/context/TestContext.java
@@ -20,7 +20,7 @@ import io.flamingock.internal.common.core.audit.AuditWriter;
 import io.flamingock.internal.core.builder.BuilderAccessor;
 import io.flamingock.internal.core.runner.Runner;
 import io.flamingock.support.domain.AuditEntryDefinition;
-import io.flamingock.support.inmemory.InMemoryAuditStore;
+import io.flamingock.support.InMemoryAuditStore;
 
 import java.util.ArrayList;
 import java.util.Collections;
